[TECH-293] Pull deploy set discovery to local
branch: feature/TECH-293-fix-step-interdependency
